Always emit change events
`<FreeformInput>` should emit change events, even if the contents
haven't changed. I hid the case where value was the same even after the
user finished editing, but that forces developers to have duplicate
keydown/blur events if they want to detect when the user is done editing
in that particular case. The event is critical when the UI shows/hides
the input depending on some other action, as is the case with
a particular app I'm building.

Closes #4